Hotfix
Changed Type of LoveValue inside IntroductionOptionSO from float to int

+++ YGFIL/Assets/_Project/Resources/Scriptable Objects/Minigames/Minigame 2/IntroductionOptionSO.cs
@@ -10,6 +10,6 @@ namespace YGFIL.ScriptableObjects
         public Image Image { get; set; }
 
         [field: SerializeField]
-        public float LoveValue { get; set; }
+        public int LoveValue { get; set; }
     }
 }
